优化

+++ mybatis-generator-demo/src/main/java/com/lee9213/mybatis/generator/config/sql/converts/MySqlTypeConvert.java
@@ -22,7 +22,7 @@ public class MySqlTypeConvert implements ITypeConvert {
         } else if (t.contains("tinyint(1)")) {
             return DbColumnType.BOOLEAN;
         }else if (t.contains("tinyint(4)")) {
-            return DbColumnType.BYTE;
+            return DbColumnType.INTEGER;
         } else if (t.contains("int")) {
             return DbColumnType.INTEGER;
         } else if (t.contains("text")) {
